requirements.txt
scapy
whois
dnspython
ipwhois
requests
ipwhois
streamlit
plotly
pandas
